+++ aldrin/src/discoverer/test.rs
@@ -17,14 +17,14 @@ async fn any_object_no_services() {
     let obj = client.create_object(ObjectUuid::new_v4()).await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
 
     obj.destroy().await.unwrap();
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -49,14 +49,14 @@ async fn specific_object_no_services() {
         .unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
 
     obj.destroy().await.unwrap();
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -82,7 +82,7 @@ async fn specific_object() {
         .unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -91,7 +91,7 @@ async fn specific_object() {
 
     obj.destroy().await.unwrap();
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -117,7 +117,7 @@ async fn any_object() {
         .unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -126,7 +126,7 @@ async fn any_object() {
 
     obj.destroy().await.unwrap();
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -136,7 +136,7 @@ async fn any_object() {
     let svc2 = obj.create_service(svc2.id().uuid, 0).await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -145,7 +145,7 @@ async fn any_object() {
 
     obj.destroy().await.unwrap();
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -160,7 +160,7 @@ async fn empty() {
     let mut client = broker.add_client().await;
 
     let mut discoverer = client.create_discoverer::<()>().build().await.unwrap();
-    assert!(discoverer.next_event().await.is_none());
+    assert!(discoverer.next_event_ref().await.is_none());
     assert!(discoverer.is_finished());
 
     client.join().await;
@@ -188,14 +188,14 @@ async fn current_only() {
     let _svc2 = obj.create_service(ServiceUuid::new_v4(), 0).await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
     assert_eq!(ev.service_id(svc1.id().uuid), svc1.id());
     assert_eq!(ev.service_id(svc2.id().uuid), svc2.id());
 
-    assert!(discoverer.next_event().await.is_none());
+    assert!(discoverer.next_event_ref().await.is_none());
     assert!(discoverer.is_finished());
 
     client.join().await;
@@ -218,7 +218,7 @@ async fn restart_any() {
         .unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -227,7 +227,7 @@ async fn restart_any() {
     discoverer.restart().await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -253,7 +253,7 @@ async fn restart_specific() {
         .unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -262,7 +262,7 @@ async fn restart_specific() {
     discoverer.restart().await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -290,14 +290,14 @@ async fn restart_current_only_any() {
     let svc = obj.create_service(uuid, 0).await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await;
+    let ev = discoverer.next_event_ref().await;
     assert!(ev.is_none());
     assert!(discoverer.is_finished());
 
     discoverer.restart_current_only().await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
@@ -326,14 +326,14 @@ async fn restart_current_only_specific() {
     let svc = obj.create_service(svc_uuid, 0).await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await;
+    let ev = discoverer.next_event_ref().await;
     assert!(ev.is_none());
     assert!(discoverer.is_finished());
 
     discoverer.restart_current_only().await.unwrap();
 
     assert!(!discoverer.is_finished());
-    let ev = discoverer.next_event().await.unwrap();
+    let ev = discoverer.next_event_ref().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());